db処理系以外はコントローラで行う
