+++ server/app/controllers/webserver.go
@@ -140,11 +140,44 @@ func (a *App) userUpdateHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	a.DB.UpdateUser(foundUser, r)
-	if err := a.DB.UpdateUser(foundUser, r); err != nil {
-		a.apiError(w, "success", http.StatusOK)
+	type body struct {
+		User struct {
+			Password string `json:"password,omitempty"`
+		} `json:"user,omitempty"`
+		NewUser struct {
+			UserName string `json:"user_name,omitempty"`
+			Email    string `json:"email,omitempty"`
+			Password string `json:"password,omitempty"`
+		} `json:"new_user,omitempty"`
+	}
+
+	var updateUser body
+	json.NewDecoder(r.Body).Decode(&updateUser)
+
+	if err := bcrypt.CompareHashAndPassword([]byte(foundUser.Password), []byte(updateUser.User.Password)); err != nil {
+		a.apiError(w, err.Error(), http.StatusNotAcceptable)
 		return
 	}
+	if updateUser.NewUser.UserName != "" {
+		foundUser.UserName = updateUser.NewUser.UserName
+	}
+	if updateUser.NewUser.Email != "" {
+		foundUser.Email = updateUser.NewUser.Email
+	}
+	if updateUser.NewUser.Password != "" {
+		hash, err := bcrypt.GenerateFromPassword([]byte(updateUser.NewUser.Password), 10)
+		if err != nil {
+			a.apiError(w, err.Error(), http.StatusBadRequest)
+			return
+		}
+		foundUser.Password = string(hash)
+	}
+
+	if err := a.DB.UpdateUser(foundUser); err != nil {
+		a.apiError(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
 	a.apiError(w, "success", http.StatusOK)
 	return
 }
