+++ server/app/entity/db.go
@@ -1,10 +1,8 @@
 package entity
 
-import "net/http"
-
 type DB interface {
 	CreateUser(name, email, pass string) (err error)
 	FindUserByID(id int) (User, error)
-	UpdateUser(foundUser User, r *http.Request) (err error)
+	UpdateUser(foundUser User) (err error)
 	DeleteUser(id int, pass string) (err error)
 }
