+++ server/app/models/user.go
@@ -1,9 +1,7 @@
 package models
 
 import (
-	"encoding/json"
 	"log"
-	"net/http"
 	"time"
 
 	"index-indicator-apis/server/app/entity"
@@ -58,42 +56,9 @@ func (m *Models) FindUserByID(id int) (entity.User, error) {
 }
 
 // UpdateUser userアップデート
-func (m *Models) UpdateUser(foundUser entity.User, r *http.Request) (err error) {
-	type body struct {
-		User struct {
-			Password string `json:"password,omitempty"`
-		} `json:"user,omitempty"`
-		NewUser struct {
-			UserName string `json:"user_name,omitempty"`
-			Email    string `json:"email,omitempty"`
-			Password string `json:"password,omitempty"`
-		} `json:"new_user,omitempty"`
-	}
-
-	var updateUser body
-	json.NewDecoder(r.Body).Decode(&updateUser)
-
-	if err := bcrypt.CompareHashAndPassword([]byte(foundUser.Password), []byte(updateUser.User.Password)); err != nil {
-		return err
-	}
-
-	if updateUser.NewUser.UserName != "" {
-		foundUser.UserName = updateUser.NewUser.UserName
-	}
-	if updateUser.NewUser.Email != "" {
-		foundUser.Email = updateUser.NewUser.Email
-	}
-	if updateUser.NewUser.Password != "" {
-		hash, err := bcrypt.GenerateFromPassword([]byte(updateUser.NewUser.Password), 10)
-		if err != nil {
-			log.Fatal(err)
-		}
-		foundUser.Password = string(hash)
-	}
-
+func (m *Models) UpdateUser(foundUser entity.User) (err error) {
 	if err := m.DB.Save(&foundUser).Error; err != nil {
 		return err
 	}
-
 	return nil
 }
